[Refactor] Modularize SCM CI inspection clients (#701)
