+++ packages/ado/src/api.ts
@@ -18,13 +18,12 @@ import {
   resolveDeploymentEnvVar,
 } from '@roomote/db/server';
 
+export * from './ci';
+
 const ADO_PROVIDER = 'ado' satisfies SourceControlProvider;
 const DEFAULT_ADO_BASE_URL = 'https://dev.azure.com';
-const ADO_API_VERSION = '7.1';
+export const ADO_API_VERSION = '7.1';
 const ADO_TOKEN_VALIDATION_TIMEOUT_MS = 10_000;
-const ADO_FAILURE_EVIDENCE_MAX_TASKS = 5;
-const ADO_FAILURE_EVIDENCE_TRACE_CHARS = 6_000;
-const ADO_BUILD_INSPECTION_TIMEOUT_MS = 5_000;
 const ADO_ENTRA_TOKEN_SCOPE = 'https://app.vssps.visualstudio.com/.default';
 const ADO_ENTRA_RESOURCE_SCOPE =
   '499b84ac-1321-427f-aa17-267ca6975798/.default';
@@ -439,7 +438,7 @@ function isEntraAccessToken(token: string): boolean {
   return token.split('.').length === 3;
 }
 
-function buildAdoAuthorizationHeader(token: string): string {
+export function buildAdoAuthorizationHeader(token: string): string {
   if (isEntraAccessToken(token)) {
     return `Bearer ${token}`;
   }
@@ -476,7 +475,7 @@ export function buildAdoOrganizationApiBaseUrl({
   );
 }
 
-function buildAdoApiUrl(
+export function buildAdoApiUrl(
   organizationApiBaseUrl: string,
   path: string,
   params: Record<string, string | number | boolean>,
@@ -541,7 +540,7 @@ async function requestAdoJson<T>({
   };
 }
 
-function parseAdoRepositoryFullName(repositoryFullName: string): {
+export function parseAdoRepositoryFullName(repositoryFullName: string): {
   organization: string;
   project: string;
   repository: string;
@@ -558,7 +557,7 @@ function parseAdoRepositoryFullName(repositoryFullName: string): {
   return { organization, project, repository };
 }
 
-async function resolveAdoOrganizationApiBaseUrl({
+export async function resolveAdoOrganizationApiBaseUrl({
   organization,
   baseUrl,
   organizationApiBaseUrl,
@@ -1488,477 +1487,6 @@ export function stripAdoGitRef(refName: string | null | undefined): string {
  */
 export async function resolveAdoInstanceHost(): Promise<string> {
   return hostFromBaseUrl(await resolveAdoBaseUrl()).toLowerCase();
-}
-
-const adoBuildSchema = z
-  .object({
-    id: z.number(),
-    buildNumber: z.string().optional(),
-    status: z.string().optional(),
-    result: z.string().nullable().optional(),
-    sourceBranch: z.string().optional(),
-    sourceVersion: z.string().optional(),
-    reason: z.string().optional(),
-    url: z.string().optional(),
-    definition: z
-      .object({
-        id: z.number().optional(),
-        name: z.string().optional(),
-      })
-      .passthrough()
-      .optional(),
-    project: z
-      .object({
-        id: z.string().optional(),
-        name: z.string().optional(),
-      })
-      .passthrough()
-      .optional(),
-    repository: z
-      .object({
-        id: z.string().optional(),
-        name: z.string().optional(),
-        type: z.string().optional(),
-      })
-      .passthrough()
-      .optional(),
-    _links: z
-      .object({
-        web: z
-          .object({
-            href: z.string().optional(),
-          })
-          .passthrough()
-          .optional(),
-      })
-      .passthrough()
-      .optional(),
-  })
-  .passthrough();
-
-const adoBuildListResponseSchema = z.object({
-  count: z.number().optional(),
-  value: z.array(adoBuildSchema),
-});
-
-const adoTimelineRecordSchema = z
-  .object({
-    id: z.string().optional(),
-    parentId: z.string().nullable().optional(),
-    name: z.string().optional(),
-    type: z.string().optional(),
-    result: z.string().nullable().optional(),
-    state: z.string().optional(),
-    // Real timelines send `"log": null` on records without logs; a plain
-    // .optional() rejects that and fails the whole timeline parse.
-    log: z
-      .object({
-        id: z.number().optional(),
-        url: z.string().optional(),
-      })
-      .passthrough()
-      .nullish(),
-    issues: z
-      .array(
-        z
-          .object({
-            type: z.string().optional(),
-            category: z.string().optional(),
-            message: z.string().optional(),
-          })
-          .passthrough(),
-      )
-      .nullish(),
-  })
-  .passthrough();
-
-const adoTimelineSchema = z
-  .object({
-    records: z.array(adoTimelineRecordSchema).optional(),
-  })
-  .passthrough();
-
-export type AdoBuild = z.infer<typeof adoBuildSchema>;
-type AdoTimelineRecord = z.infer<typeof adoTimelineRecordSchema>;
-
-function timelineRecordParentId(record: AdoTimelineRecord): string | undefined {
-  const parentId = record.parentId;
-  return typeof parentId === 'string' && parentId.trim() ? parentId : undefined;
-}
-
-/**
- * Keep each independent innermost failed timeline record. Cascaded Stage /
- * Phase / Job wrappers that only exist because a child task failed are
- * dropped, while a peer job that failed without a failed task child is kept.
- */
-export function selectInnermostFailedAdoTimelineRecords(
-  records: AdoTimelineRecord[],
-): AdoTimelineRecord[] {
-  const failed = records.filter(
-    (record) => (record.result ?? '').toLowerCase() === 'failed',
-  );
-  if (failed.length === 0) {
-    return [];
-  }
-
-  const allById = new Map(
-    records
-      .filter((record): record is AdoTimelineRecord & { id: string } =>
-        Boolean(record.id?.trim()),
-      )
-      .map((record) => [record.id, record]),
-  );
-
-  const hasParentLinks = failed.some((record) =>
-    Boolean(timelineRecordParentId(record)),
-  );
-
-  if (hasParentLinks) {
-    const hasFailedDescendant = (record: AdoTimelineRecord): boolean => {
-      if (!record.id) {
-        return false;
-      }
-      for (const other of failed) {
-        if (other === record) {
-          continue;
-        }
-        let cursor = timelineRecordParentId(other);
-        const seen = new Set<string>();
-        while (cursor && !seen.has(cursor)) {
-          if (cursor === record.id) {
-            return true;
-          }
-          seen.add(cursor);
-          cursor = timelineRecordParentId(allById.get(cursor) ?? {});
-        }
-      }
-      return false;
-    };
-
-    const leaves = failed.filter((record) => !hasFailedDescendant(record));
-    const typedLeaves = leaves.filter((record) => {
-      const type = (record.type ?? '').toLowerCase();
-      return type === 'task' || type === 'job' || Boolean(record.name?.trim());
-    });
-    return (typedLeaves.length > 0 ? typedLeaves : leaves).slice(
-      0,
-      ADO_FAILURE_EVIDENCE_MAX_TASKS,
-    );
-  }
-
-  // Timelines without parent links only support global depth preference.
-  const failedOfType = (type: string) =>
-    failed.filter((record) => (record.type ?? '').toLowerCase() === type);
-  const failedTasks = failedOfType('task');
-  const failedJobs = failedOfType('job');
-  const failedNamed = failed.filter((record) => Boolean(record.name?.trim()));
-  return (
-    failedTasks.length > 0
-      ? failedTasks
-      : failedJobs.length > 0
-        ? failedJobs
-        : failedNamed
-  ).slice(0, ADO_FAILURE_EVIDENCE_MAX_TASKS);
-}
-
-function buildAdoBranchRef(branch: string): string {
-  const trimmed = branch.trim();
-  if (!trimmed) {
-    return 'refs/heads/main';
-  }
-  return trimmed.startsWith('refs/') ? trimmed : `refs/heads/${trimmed}`;
-}
-
-export function getAdoBuildWebUrl(build: AdoBuild): string {
-  const web = build._links?.web?.href?.trim();
-  if (web) {
-    return web;
-  }
-  if (build.url?.trim()) {
-    return build.url.trim();
-  }
-  return `build/${build.id}`;
-}
-
-/**
- * Newest completed build for a Git repository on the given branch.
- * Does not filter by result so callers can detect already-green tips before
- * launching triage. Uses the deployment Azure DevOps credential. Null when
- * none match.
- */
-export async function getLatestAdoBuild(params: {
-  repositoryFullName: string;
-  repositoryId: string;
-  branch: string;
-  token?: string;
-  organization?: string;
-  baseUrl?: string;
-  organizationApiBaseUrl?: string;
-  fetchImpl?: typeof fetch;
-}): Promise<AdoBuild | null> {
-  const adoToken = params.token ?? (await resolveAdoToken());
-  if (!adoToken?.trim()) {
-    throw new Error('ADO_TOKEN is required to inspect Azure DevOps builds.');
-  }
-
-  const parsed = parseAdoRepositoryFullName(params.repositoryFullName);
-  const organizationApiBaseUrl = await resolveAdoOrganizationApiBaseUrl({
-    organization: params.organization ?? parsed.organization,
-    baseUrl: params.baseUrl,
-    organizationApiBaseUrl: params.organizationApiBaseUrl,
-  });
-
-  if (!organizationApiBaseUrl) {
-    throw new Error(
-      'ADO_ORGANIZATION is required to inspect Azure DevOps builds.',
-    );
-  }
-
-  const projectPath = encodeURIComponent(parsed.project);
-  const response = await (params.fetchImpl ?? fetch)(
-    buildAdoApiUrl(
-      organizationApiBaseUrl,
-      `/${projectPath}/_apis/build/builds`,
-      {
-        'api-version': ADO_API_VERSION,
-        repositoryId: params.repositoryId,
-        repositoryType: 'TfsGit',
-        branchName: buildAdoBranchRef(params.branch),
-        statusFilter: 'completed',
-        queryOrder: 'finishTimeDescending',
-        $top: 1,
-      },
-    ),
-    {
-      method: 'GET',
-      headers: {
-        Accept: 'application/json',
-        Authorization: buildAdoAuthorizationHeader(adoToken),
-      },
-      signal: AbortSignal.timeout(ADO_BUILD_INSPECTION_TIMEOUT_MS),
-    },
-  );
-
-  // Azure DevOps answers rejected PATs with a 203 sign-in page instead of a
-  // 401. Throw so callers surface a credential problem instead of reading it
-  // as "no failed build"; only an unknown project/repository maps to null.
-  if ([203, 401, 403].includes(response.status)) {
-    throw new Error(
-      `Azure DevOps rejected the access token when listing builds (status ${response.status}). Confirm it is active, belongs to the organization, and has Build read access.`,
-    );
-  }
-
-  if (response.status === 404) {
-    return null;
-  }
-
-  if (response.status !== 200) {
-    throw new Error(
-      `Azure DevOps API request failed: ${response.status} ${response.statusText}`,
-    );
-  }
-
-  const { value } = adoBuildListResponseSchema.parse(await response.json());
-  return value[0] ?? null;
-}
-
-async function readResponseTextTail(
-  response: Response,
-  maxChars: number,
-): Promise<{ text: string; truncated: boolean }> {
-  if (!response.body) {
-    const text = await response.text();
-    return {
-      text: text.slice(-maxChars),
-      truncated: text.length > maxChars,
-    };
-  }
-
-  const reader = response.body.getReader();
-  const decoder = new TextDecoder();
-  let tail = '';
-  let totalChars = 0;
-
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) {
-      break;
-    }
-
-    const chunk = decoder.decode(value, { stream: true });
-    totalChars += chunk.length;
-    tail = `${tail}${chunk}`.slice(-maxChars);
-  }
-
-  const finalChunk = decoder.decode();
-  totalChars += finalChunk.length;
-  tail = `${tail}${finalChunk}`.slice(-maxChars);
-
-  return { text: tail, truncated: totalChars > maxChars };
-}
-
-function formatAdoTimelineRecordEvidence(params: {
-  record: z.infer<typeof adoTimelineRecordSchema>;
-  logText: string | null;
-  logTruncated: boolean;
-}): string {
-  const issues = (params.record.issues ?? [])
-    .map((issue) => issue.message?.trim())
-    .filter((message): message is string => Boolean(message));
-  const metadata = [
-    `task=${JSON.stringify(params.record.name ?? 'unknown')}`,
-    ...(params.record.type
-      ? [`type=${JSON.stringify(params.record.type)}`]
-      : []),
-    ...(params.record.result
-      ? [`result=${JSON.stringify(params.record.result)}`]
-      : []),
-  ].join(' ');
-
-  const issueBlock =
-    issues.length > 0
-      ? `\nIssues:\n${issues.map((message) => `- ${message}`).join('\n')}`
-      : '';
-
-  if (!params.logText) {
-    return `${metadata}${issueBlock}${issueBlock ? '' : '\nLog trace unavailable.'}`;
-  }
-
-  return `${metadata}${issueBlock}\n${
-    params.logTruncated
-      ? '[Earlier log output omitted; showing the tail.]\n'
-      : ''
-  }${params.logText}`;
-}
-
-/**
- * Bounded, prompt-ready snapshot of failed build timeline tasks and log tails.
- */
-export async function getAdoBuildFailureEvidence(params: {
-  repositoryFullName: string;
-  buildId: number;
-  token?: string;
-  organization?: string;
-  baseUrl?: string;
-  organizationApiBaseUrl?: string;
-  fetchImpl?: typeof fetch;
-}): Promise<string | null> {
-  const adoToken = params.token ?? (await resolveAdoToken());
-  if (!adoToken?.trim()) {
-    throw new Error(
-      'ADO_TOKEN is required to inspect Azure DevOps build logs.',
-    );
-  }
-
-  const parsed = parseAdoRepositoryFullName(params.repositoryFullName);
-  const organizationApiBaseUrl = await resolveAdoOrganizationApiBaseUrl({
-    organization: params.organization ?? parsed.organization,
-    baseUrl: params.baseUrl,
-    organizationApiBaseUrl: params.organizationApiBaseUrl,
-  });
-
-  if (!organizationApiBaseUrl) {
-    throw new Error(
-      'ADO_ORGANIZATION is required to inspect Azure DevOps build logs.',
-    );
-  }
-
-  const fetchImpl = params.fetchImpl ?? fetch;
-  const projectPath = encodeURIComponent(parsed.project);
-  const timelineResponse = await fetchImpl(
-    buildAdoApiUrl(
-      organizationApiBaseUrl,
-      `/${projectPath}/_apis/build/builds/${params.buildId}/timeline`,
-      { 'api-version': ADO_API_VERSION },
-    ),
-    {
-      method: 'GET',
-      headers: {
-        Accept: 'application/json',
-        Authorization: buildAdoAuthorizationHeader(adoToken),
-      },
-      signal: AbortSignal.timeout(ADO_BUILD_INSPECTION_TIMEOUT_MS),
-    },
-  );
-
-  if ([203, 401, 403, 404].includes(timelineResponse.status)) {
-    return null;
-  }
-
-  if (timelineResponse.status !== 200) {
-    throw new Error(
-      `Azure DevOps API request failed: ${timelineResponse.status} ${timelineResponse.statusText}`,
-    );
-  }
-
-  const timeline = adoTimelineSchema.parse(await timelineResponse.json());
-  // Cascaded Stage/Phase/Job wrappers share parent links with their failed
-  // children; keep each independent innermost failure so peer jobs are not
-  // dropped when another job already has a failed task.
-  const failedRecords = selectInnermostFailedAdoTimelineRecords(
-    timeline.records ?? [],
-  );
-
-  if (failedRecords.length === 0) {
-    return null;
-  }
-
-  const evidence = await Promise.all(
-    failedRecords.map(async (record) => {
-      const logId = record.log?.id;
-      if (logId === undefined) {
-        return formatAdoTimelineRecordEvidence({
-          record,
-          logText: null,
-          logTruncated: false,
-        });
-      }
-
-      try {
-        const logResponse = await fetchImpl(
-          buildAdoApiUrl(
-            organizationApiBaseUrl,
-            `/${projectPath}/_apis/build/builds/${params.buildId}/logs/${logId}`,
-            { 'api-version': ADO_API_VERSION },
-          ),
-          {
-            method: 'GET',
-            headers: {
-              Accept: 'text/plain',
-              Authorization: buildAdoAuthorizationHeader(adoToken),
-            },
-            signal: AbortSignal.timeout(ADO_BUILD_INSPECTION_TIMEOUT_MS),
-          },
-        );
-
-        if (logResponse.status !== 200) {
-          return formatAdoTimelineRecordEvidence({
-            record,
-            logText: null,
-            logTruncated: false,
-          });
-        }
-
-        const tail = await readResponseTextTail(
-          logResponse,
-          ADO_FAILURE_EVIDENCE_TRACE_CHARS,
-        );
-        return formatAdoTimelineRecordEvidence({
-          record,
-          logText: tail.text.trim() || null,
-          logTruncated: tail.truncated,
-        });
-      } catch {
-        return formatAdoTimelineRecordEvidence({
-          record,
-          logText: null,
-          logTruncated: false,
-        });
-      }
-    }),
-  );
-
-  return evidence.join('\n\n');
 }
 
 export type AdoServiceHookRemoveResult = {
