+++ packages/bitbucket/src/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
+export * from './ci';
 export * from './oauth';
